chore: move components to separate file
